[neptune/#113] Field Visit workflow - Maintenance Adds update model to FieldVisit/EditMaintenanceRecord. Tries to be more "delete-full" about Field Visits
